fix office

+++ resolvers/office/deleteOffice.js
@@ -1,18 +1,34 @@
-const { Office } = require("../../models")
-const { authentication } = require("../../helpers/authentication")
-const { authorizationIdOfficeUser } = require("../../helpers/authorizationIdOfficeUser")
-module.exports = authentication( authorizationIdOfficeUser( async (_,args) => {
-    try {
-        const officeId = args.id
+const { Office } = require("../../models");
+const { authentication } = require("../../helpers/authentication");
+const {
+  authorizationUserRole,
+} = require("../../helpers/authorizationUserRole");
+
+module.exports = authentication(
+  authorizationUserRole(
+    ["pengepul", "pengrajin"],
+    async (_, args, { user }) => {
+      try {
+        const currentOffice = await Office.findOne({
+          where: { UserId: user.id },
+        });
+        if (!currentOffice) {
+          throw new Error("office not found");
+        }
         const deletedOffice = await Office.destroy({
-            where : {
-                id : officeId
-            },
-        })
-        let message = null
-        deletedOffice ? message = {msg : "success delete office data"} : message = {msg : "data not found"}
-        return message
-    } catch (error) {
-        console.log(error)
+          where: {
+            id: currentOffice.id,
+          },
+        });
+        let message = null;
+        deletedOffice
+          ? (message = { msg: "success delete office data" })
+          : (message = { msg: "data not found" });
+        return message;
+      } catch (error) {
+        return error;
+      }
     }
-}))
+  )
+);
+
